Finish draft of first section with more flexbox
I finished adding some more flexbox. For all simplified intents and
purposes, the page now 90% matches the first section of the example
image.

Added flexbox to more pieces, added an image of a dog, figured out the
height and width of said dog. Next stop is messing with the next
section. This section should probably be more polished before I'm
done with it for real.