Add content and styling to second section
Added the basic content - photos of birds and some text about them -
and styling to the second section, the area just below the header.
Don't forget to add credit for these images to the README.md. They are
currently credited in the alt text and there's a reminder in the html
comments.